total component props using redux
