+++ src/components/Total.js
@@ -1,6 +1,7 @@
-import React from 'react';
+import React from "react";
+import { connect } from "react-redux";
 
-const Total = props => {
+export const Total = props => {
   return (
     <div className="content">
       <h4>Total Amount: ${props.car.price + props.additionalPrice}</h4>
@@ -8,4 +9,11 @@ const Total = props => {
   );
 };
 
-export default Total;
+const mapStateToProps = state => {
+  return {
+    ...state,
+    car: { ...state.car, price: state.car.price + state.additionalPrice }
+  };
+};
+
+export default connect(mapStateToProps)(Total);
