+++ src/reducers/index.js
@@ -6,7 +6,13 @@ export const initialState = {
     image:
       "https://cdn.motor1.com/images/mgl/0AN2V/s1/2019-ford-mustang-bullitt.jpg",
     features: []
-  }
+  },
+  additionalFeatures: [
+    { id: 1, name: "V-6 engine", price: 1500 },
+    { id: 2, name: "Racing detail package", price: 1500 },
+    { id: 3, name: "Premium sound system", price: 500 },
+    { id: 4, name: "Rear spoiler", price: 250 }
+  ]
 };
 
 export const reducer = (state = initialState, action) => {
